Added optional AVIF support
Not enabled by default because it requires a lot of external dependencies and in the end dynamically links to the dav1d.dll leaking an external dependency to after build.

Note that the image implementation is not handling a decode error, so even if you get this running it will not work at the time of this commit.