Fixed issue when the bidi override width of a line ends-up returned as the measured width for the line in a subsequent layout cycle causing incorrect wraps.
The fix consists of saving the initial size "original" and used that size for measure return info.

The issue was caused by a full "reshape" not getting triggered in text when two measure calls happens with the same inline constrains, and a "reshape_lines" happens in between, the "reshape_lines" applies the override size, and the secound measure returns it instead of the original measure size., this inccorect size affects the available size of the next text widgets that get wrap incorrectly.