Fix style property order.
When we set `inner; outer;` in a widget we build to `outer(inner)`, but
if `inner` was in a style and `outer` on the widget we where building
`inner(outer)` because we where extending the widget builder with the
style builder. Property override was working only due to the
`Importance` value, this issue becomes visible when using multiple
different background fill properties.