Fix config errors (#115)
First, the errors are only logged, enabled test_log so now config tests fail for log errors.

The issue was caused by conversions of number types between TomlValue and YamlValue. Serde JSON does not implicit cast
floats to integers, and we converted to float first. Changing the order of conversion fixed the issue.