Fix prebuild build script not running in checked out repo
Issue is after merging the release only the Cargo.toml updates, not the build script, and because we use rerun-if-changed all other project files must be manually included as well.

Also try to rerun in case of network error, not sure if this will work. Depends on when cargo caches the file timestamps.