Hotfix `zng-wgt-scroll@0.6.4`
This crate depends on `zng-wgt -> zng-layout -> zng-unit`.
A later release of `zng-unit` implements `Ord for Factor`
and deprecates the custom `Factor::clamp`. Unfortunately Ord
implements for all `&mut A: Ord` too, and it is on the Rust prelude
so it ends up shadowing the custom `clamp`,
breaking the build of `zng-wgt-scroll`.

The hotfix will be released manually to fix this,
without a full project release.