Replace "bind ref" methods with "bind modify" methods.
MapRef used more memory the a simple deref should, and it did not avoid the need for contextual wrapping anyway. "bind modify" style methods are more powerful, for example the CONFIG backends can be implemented using it, so map_ref would end-up as a third mapping option in Var that uses would have to consider, when it does not even give a measurable performance gain.