Refactor Var API (#702)
* New var API

The primary motivation for this rewrite is reduction of generics code bloat,
the variable types are unified to the new struct `Var<T>`. The general shape
of the API remains the same.

* Replace zng-var with new API

* Add missing code

* Upgrade var API uses

* Better names

The idea of var_ prefix does not flow when reading

* Rename var_local to const_var

The value now is only "local" if it fits in the SmallBox stack space.

* Last renames, probably

* Fixes

* Another missed rename

* More fixes

* Implement Debug for vars

* Add VarEq

* Fix contextualization of mapping

* More map contextualization  fixes

* More fixes

* More fixes

* More dbg

* More fixes and some DATA improvements

* Replace "bind ref" methods with "bind modify" methods.

MapRef used more memory the a simple deref should, and it did not avoid the need for contextual wrapping anyway. "bind modify" style methods are more powerful, for example the CONFIG backends can be implemented using it, so map_ref would end-up as a third mapping option in Var that uses would have to consider, when it does not even give a measurable performance gain.

* var_instance_tag in hook args

* TODO review

* More fixes

* Fix zero sized gradients causing render panic

* More fixes

* Fix `accepts_enter` and `accepts_tag` in text editor widgets

* More fixes.

* Fix Android build

Unrelated to the var rewrite

* Doc fixes

* For some reason these doc tests fail only when run with cargo do test --doc, not cargo do test -p zng-var --doc

* Use enum dispatch for vars