Downgrade bincode dependency (#885)
* Downgrade bincode dependency

There is some weirdness with this package and it will probably become a unmaintained security issue in the future, fow now this goes back to
version 1.3.3 that is the stable version used by `ipc-channel`, hopefully we can just follow what ipc-channel does in the future.

* Upgrade flume

Not a breaking change for us (removed feature).  New version has bug fixes.